+++ src/utopia/coder/controller/parsing/scala/InstanceBuilder.scala
@@ -2,7 +2,7 @@ package utopia.coder.controller.parsing.scala
 
 import utopia.coder.model.scala.code.{Code, CodeLine}
 import utopia.coder.model.scala.datatype.{Extension, GenericType, Reference}
-import utopia.coder.model.scala.declaration.DeclarationPrefix.{Abstract, Case, Sealed}
+import utopia.coder.model.scala.declaration.DeclarationPrefix.{Abstract, Case, Implicit, Sealed}
 import utopia.coder.model.scala.declaration.InstanceDeclarationType.{ClassD, ObjectD, TraitD}
 import utopia.coder.model.scala.declaration._
 import utopia.coder.model.scala.doc.ScalaDoc
@@ -81,7 +81,7 @@ class InstanceBuilder(visibility: Visibility, prefixes: Set[DeclarationPrefix], 
 					typesBuilder.result(), freeCode, propertiesBuilder.result(), methodsBuilder.result().toSet,
 					nestedBuilder.result().toSet, visibility, annotations, scalaDoc.description, scalaDoc.author,
 					commentsBefore, since, isCaseClass = prefixes.contains(Case),
-					isAbstract = prefixes.contains(Abstract))
+					isAbstract = prefixes.contains(Abstract), isImplicit = prefixes.contains(Implicit))
 			case TraitD =>
 				val freeCode = freeCodeBuilder.result()
 				if (freeCode.nonEmpty) {